Fixes db names.

+++ db/schema.rb
@@ -12,13 +12,13 @@
 
 ActiveRecord::Schema.define(version: 2019_12_26_010237) do
 
+  # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
-
 
   create_table "cards", force: :cascade do |t|
     t.string "name"
     t.text "content"
-    t.integer "dashboard_id", null: false
+    t.bigint "dashboard_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["dashboard_id"], name: "index_cards_on_dashboard_id"
@@ -26,7 +26,7 @@ ActiveRecord::Schema.define(version: 2019_12_26_010237) do
 
   create_table "dashboards", force: :cascade do |t|
     t.string "name"
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.index ["user_id"], name: "index_dashboards_on_user_id"
